episode type support

+++ esp/spotte-esp/screen.ino
@@ -7,8 +7,8 @@
 
 Adafruit_SH1106G display = Adafruit_SH1106G(SCREEN_WIDTH, SCREEN_HEIGHT, &Wire, SCREEN_RESET);
 
-char* parseTimeValue(int ms) {
-  long totalSeconds = round(ms / 1000);
+char* parseTimeValue(unsigned long ms) {
+  unsigned long totalSeconds = round(ms / 1000);
   static char output[8];
   int hours = totalSeconds / 3600;
   int minutes = (totalSeconds % 3600) / 60;
